fix(smoke): restore fresh-host harness
Preserve authenticated state while initializing validator-specific node state, acknowledge the controlled-devnet backend, and make the disk-space calculation portable to macOS.

+++ crates/node/src/storage_migration.rs
@@ -984,9 +984,8 @@ fn available_disk_bytes(path: &Path) -> io::Result<u64> {
     }
     // SAFETY: statvfs returned success and initialized the output structure.
     let stats = unsafe { stats.assume_init() };
-    stats
-        .f_bavail
-        .checked_mul(stats.f_frsize)
+    u64::from(stats.f_bavail)
+        .checked_mul(u64::from(stats.f_frsize))
         .ok_or_else(|| io::Error::new(io::ErrorKind::InvalidData, "available disk overflow"))
 }
